+++ Backend/routers/diary.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const db = require("../db");
+const db = require("../index.js");
 const diaryRouter = express.Router();
